Separate lazy concatenate + rename scil_streamlines_math
