+++ scripts/tests/test_tractogram_math.py
@@ -12,7 +12,7 @@ tmp_dir = tempfile.TemporaryDirectory()
 
 
 def test_help_option(script_runner):
-    ret = script_runner.run('scil_streamlines_math.py', '--help')
+    ret = script_runner.run('scil_tractogram_math.py', '--help')
     assert ret.success
 
 
@@ -22,7 +22,7 @@ def test_execution_lazy_concatenate_no_color(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'lazy_concatenate',
+    ret = script_runner.run('scil_tractogram_math.py', 'lazy_concatenate',
                             in_tracto_1, in_tracto_2,
                             'lazy_concatenate.trk')
     assert ret.success
@@ -34,7 +34,7 @@ def test_execution_lazy_concatenate_mix(script_runner):
                                'fibercup_bundles_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'lazy_concatenate',
+    ret = script_runner.run('scil_tractogram_math.py', 'lazy_concatenate',
                             in_tracto_1, in_tracto_2,
                             'lazy_concatenate_mix.trk')
     assert ret.success
@@ -46,7 +46,7 @@ def test_execution_union_no_color(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'union',
+    ret = script_runner.run('scil_tractogram_math.py', 'union',
                             in_tracto_1, in_tracto_2, 'union.trk')
     assert ret.success
 
@@ -57,7 +57,7 @@ def test_execution_intersection_no_color(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'intersection',
+    ret = script_runner.run('scil_tractogram_math.py', 'intersection',
                             in_tracto_1, in_tracto_2, 'intersection.trk')
     assert ret.success
 
@@ -68,7 +68,7 @@ def test_execution_difference_no_color(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'difference',
+    ret = script_runner.run('scil_tractogram_math.py', 'difference',
                             in_tracto_1, in_tracto_2, 'difference.trk')
     assert ret.success
 
@@ -79,7 +79,7 @@ def test_execution_concatenate_no_color(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'concatenate',
+    ret = script_runner.run('scil_tractogram_math.py', 'concatenate',
                             in_tracto_1, in_tracto_2, 'concatenate.trk')
     assert ret.success
 
@@ -90,7 +90,7 @@ def test_execution_union_no_color_robust(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'union',
+    ret = script_runner.run('scil_tractogram_math.py', 'union',
                             in_tracto_1, in_tracto_2, 'union_r.trk',
                             '--robust')
     assert ret.success
@@ -102,7 +102,7 @@ def test_execution_intersection_no_color_robust(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'intersection',
+    ret = script_runner.run('scil_tractogram_math.py', 'intersection',
                             in_tracto_1, in_tracto_2, 'intersection_r.trk',
                             '--robust')
     assert ret.success
@@ -114,7 +114,7 @@ def test_execution_difference_no_color_robust(script_runner):
                                'fibercup_bundles.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'difference',
+    ret = script_runner.run('scil_tractogram_math.py', 'difference',
                             in_tracto_1, in_tracto_2, 'difference_r.trk',
                             '--robust')
     assert ret.success
@@ -126,7 +126,7 @@ def test_execution_union_color(script_runner):
                                'fibercup_bundles_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0_color.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'union',
+    ret = script_runner.run('scil_tractogram_math.py', 'union',
                             in_tracto_1, in_tracto_2, 'union_color.trk')
     assert ret.success
 
@@ -137,7 +137,7 @@ def test_execution_intersection_color(script_runner):
                                'fibercup_bundles_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0_color.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'intersection',
+    ret = script_runner.run('scil_tractogram_math.py', 'intersection',
                             in_tracto_1, in_tracto_2, 'intersection_color.trk')
     assert ret.success
 
@@ -148,7 +148,7 @@ def test_execution_difference_color(script_runner):
                                'fibercup_bundles_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0_color.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'difference',
+    ret = script_runner.run('scil_tractogram_math.py', 'difference',
                             in_tracto_1, in_tracto_2, 'difference_color.trk')
     assert ret.success
 
@@ -159,7 +159,7 @@ def test_execution_concatenate_color(script_runner):
                                'fibercup_bundles_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0_color.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'concatenate',
+    ret = script_runner.run('scil_tractogram_math.py', 'concatenate',
                             in_tracto_1, in_tracto_2, 'concatenate_color.trk')
     assert ret.success
 
@@ -171,7 +171,7 @@ def test_execution_union_mix(script_runner):
                                'fibercup_bundles_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'union',
+    ret = script_runner.run('scil_tractogram_math.py', 'union',
                             in_tracto_1, in_tracto_2, 'union_mix.trk')
     assert not ret.success
 
@@ -182,7 +182,7 @@ def test_execution_intersection_mix_fake(script_runner):
                                'fibercup_bundles_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'intersection',
+    ret = script_runner.run('scil_tractogram_math.py', 'intersection',
                             in_tracto_1, in_tracto_2, 'intersection_mix.trk',
                             '--fake_metadata')
     assert ret.success
@@ -194,7 +194,7 @@ def test_execution_difference_empty_result(script_runner):
                                'fibercup_bundle_0.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0_color.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'difference',
+    ret = script_runner.run('scil_tractogram_math.py', 'difference',
                             in_tracto_1, in_tracto_2,
                             'difference_empty_results.trk', '--no_metadata')
     assert ret.success
@@ -206,7 +206,7 @@ def test_execution_difference_empty_input_1(script_runner):
                                'empty.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'fibercup_bundle_0_color.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'difference',
+    ret = script_runner.run('scil_tractogram_math.py', 'difference',
                             in_tracto_1, in_tracto_2, 'difference_empty_1.trk',
                             '--no_metadata')
     assert ret.success
@@ -218,7 +218,7 @@ def test_execution_difference_empty_input_2(script_runner):
                                'fibercup_bundle_0_color.trk')
     in_tracto_2 = os.path.join(get_home(), 'others',
                                'empty.trk')
-    ret = script_runner.run('scil_streamlines_math.py', 'difference',
+    ret = script_runner.run('scil_tractogram_math.py', 'difference',
                             in_tracto_1, in_tracto_2, 'difference_empty_2.trk',
                             '--no_metadata')
     assert ret.success
